Remove quantile calculation from `Summary`
Expose only `sum` and `count` instead, with no quantiles / percentiles.

The main reason to do this is that all this change is based on the idea
of having a "Store" interface that can be shared between Ruby processes.

The `quantile` gem doesn't play well with this, since we'd need to store
instances of `Quantile::Estimator`, which is a complex data structure, and
tricky to share between Ruby processes.

Moreover, individual Summaries on different processes cannot be aggregated,
so all processes would actually have to share one instance of this class,
which makes it extremely tricky, particularly to do performantly.

Even though this is a loss of functionality, this puts the Ruby client
on par with other client libraries, like the Python one, which also only
offers sum and count without quantiles.

Also, this is actually more compliant with the Client Library best practices:
- Summary is ENCOURAGED to offer quantiles as exports
- MUST allow not having quantiles, since just count and sum is quite useful
- This MUST be the default

The original client didn't comply with the last 2 rules, where this one
does, just like the Python client.

And quantiles, while seemingly the point of summaries, are encouraged but
not required.

I'm not discarding the idea of adding quantiles back. I have ideas on how
this could be done, but it'd probably be pretty expensive, as a single
Estimator instance would have to be marshalled / unmarshalled into bytes
for every call to `observe`, and it'd have a hard requirement of having
some sort of low-cardinality Process ID as a label (possible with Unicorn)
to avoid aggregation.