Change `text_spec.rb` to be more like an integration test
The existing spec on `Formats::Text` is based on mocking a registry with
fake metrics and values inside, which means if anything changes in the
interface for metrics, the test will not catch it.

And there's no test validating that interface.

This test is more realistic, and it actually catches the kind of bugs
we introduced in the process of this refactor

Signed-off-by: Daniel Magliola <danielmagliola@gocardless.com>