Add `Metric#with_labels` method
`with_labels` returns a metric object, augmented with some pre-set labels.
This is useful to be able to reduce repetition if a certain part of the
code is passing the same label over and over again, but the label can
have different values in different parts of the codebase (so passing in
`preset_labels` when declaring the metric is not appropriate.

Signed-off-by: Daniel Magliola <danielmagliola@gocardless.com>