Update Histogram to return buckets as strings
In the Text export format, the `le` label is reported as a string.

Moreover, some of our stores may coerce Float label values into Strings.
This is fine, since Labelsets should be a hash of symbols to strings.
It also makes sense since one of those "buckets" will be "+Inf".

It's also somewhat unavoidable, since neither the Store nor the Metric
should be coercing those back into Floats, so report the buckets as
strings instead of Floats.

Signed-off-by: Daniel Magliola <danielmagliola@gocardless.com>