Add `DataStores::DirectFileStore`
This store keeps all its data in files, one file per process and per metric.
These files have binary data mapping keys to Floats, and they're read/written
at the specific offsets of these Floats.

This is generally the recommended store to use to deal with pre-fork servers
and other "multi-process" scenarios, at least until we crack `mmap`.

This seems to be, for now, the fastest possible way to safely share data
between all the processes. Because we never actually `fsync`, we never
actually touch the disk, and FS caching makes this extremely fast. Almost
as fast as the `mmaps` without all the added risk, or the burden of the
C extensions.

Each process gets their own file for a metric. When a Prometheus scrape
is received in one of the processes, it finds all the files for a metric,
reads their values and aggregates them. We use `flock` to guarantee
consistency.

Signed-off-by: Daniel Magliola <danielmagliola@gocardless.com>