Proactively merge `base_labels`
The way `base_labels` currently work, they get set for a metric, and
all internal values for the metric are stored without those base_labels.

Then, when exporting, the Text formatter reads the internal values
without base labels, reads the base labels, and merges them.

This is not ideal for abstracting the stores away. Change this so that
the internal values always have the base labels merged in already, and
no one outside the metric itself needs to know about base_labels or deal
with them.